+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Unreleased
 
+* Add `Model#cells`
 * Add `Configuration#create_table_options`
 * Include shard in underlying index name